Renamed curve tools
