+++ gt/tools/curve_shape_to_python/curve_shape_to_python.py
