+++ gt/tools/curve_to_python/curve_to_python.py
