+++ gt/tools/package_setup/gt_tools_maya_menu.py
@@ -107,11 +107,11 @@ def load_menu(*args):
                        tooltip='Generates a single curve containing all shapes necessary to produce a word/text.',
                        icon=resource_library.Icon.tool_crv_text)
     menu.add_menu_item(label='Extract Python Curve',
-                       command=IMPORT_TOOL + 'initialize_tool("shape_curve_to_python")',
+                       command=IMPORT_TOOL + 'initialize_tool("curve_to_python")',
                        tooltip='Generates the python code necessary to create a selected curve.',
                        icon=resource_library.Icon.tool_crv_python)
     menu.add_menu_item(label='Extract Curve State',
-                       command=IMPORT_TOOL + 'initialize_tool("shape_extract_state")',
+                       command=IMPORT_TOOL + 'initialize_tool("curve_shape_to_python")',
                        tooltip='Generates the python command necessary to reshape curves back to their stored state.',
                        icon=resource_library.Icon.tool_crv_extract_state)
